adding new pbr car example
