+++ examples/CMakeLists.txt
@@ -10,7 +10,8 @@ endfunction()
 
 add_example(r3d_basic "examples/basic.c")
 add_example(r3d_lights "examples/lights.c")
-add_example(r3d_pbr "examples/pbr.c")
+add_example(r3d_pbr_musket "examples/pbr_musket.c")
+add_example(r3d_pbr_car "examples/pbr_car.c")
 add_example(r3d_transparency "examples/transparency.c")
 add_example(r3d_skybox "examples/skybox.c")
 add_example(r3d_sponza "examples/sponza.c")
